+++ mne_connectivity/spectral/__init__.py
@@ -1,3 +1,2 @@
-from .epochs_bivariate import spectral_connectivity_epochs
-from .epochs_multivariate import spectral_connectivity_epochs_multivariate
+from .epochs import spectral_connectivity_epochs
 from .time import spectral_connectivity_time
