--- mne_connectivity/spectral/epochs_bivariate.py
@@ -1,729 +1,0 @@
-# Authors: Martin Luessi <mluessi@nmr.mgh.harvard.edu>
-#          Denis A. Engemann <denis.engemann@gmail.com>
-#          Adam Li <adam2392@gmail.com>
-#          Thomas S. Binns <t.s.binns@outlook.com>
-#
-# License: BSD (3-clause)
-
-import numpy as np
-from mne.utils import logger, verbose
-
-from .epochs import (
-    _AbstractConEstBase, _check_spectral_connectivity_epochs_settings,
-    _check_spectral_connectivity_epochs_data, _get_n_epochs,
-    _prepare_connectivity, _assemble_spectral_params,
-    _compute_spectral_methods_epochs, _store_results)
-from ..utils import fill_doc, check_indices
-
-
-def _check_indices(indices, n_signals):
-    if indices is None:
-        logger.info('only using indices for lower-triangular matrix')
-        # only compute r for lower-triangular region
-        indices_use = np.tril_indices(n_signals, -1)
-    else:
-        indices_use = check_indices(indices)
-
-    # number of connectivities to compute
-    n_cons = len(indices_use[0])
-    logger.info('    computing connectivity for %d connections' % n_cons)
-
-    return n_cons, indices_use
-
-
-########################################################################
-# Bivariate connectivity estimators
-
-
-class _EpochMeanConEstBase(_AbstractConEstBase):
-    """Base class for methods that estimate connectivity as mean epoch-wise."""
-
-    patterns = None
-
-    def __init__(self, n_cons, n_freqs, n_times):
-        self.n_cons = n_cons
-        self.n_freqs = n_freqs
-        self.n_times = n_times
-
-        if n_times == 0:
-            self.csd_shape = (n_cons, n_freqs)
-        else:
-            self.csd_shape = (n_cons, n_freqs, n_times)
-
-        self.con_scores = None
-
-    def start_epoch(self):  # noqa: D401
-        """Called at the start of each epoch."""
-        pass  # for this type of con. method we don't do anything
-
-    def combine(self, other):
-        """Include con. accumated for some epochs in this estimate."""
-        self._acc += other._acc
-
-
-class _CohEstBase(_EpochMeanConEstBase):
-    """Base Estimator for Coherence, Coherency, Imag. Coherence."""
-
-    accumulate_psd = True
-
-    def __init__(self, n_cons, n_freqs, n_times):
-        super(_CohEstBase, self).__init__(n_cons, n_freqs, n_times)
-
-        # allocate space for accumulation of CSD
-        self._acc = np.zeros(self.csd_shape, dtype=np.complex128)
-
-    def accumulate(self, con_idx, csd_xy):
-        """Accumulate CSD for some connections."""
-        self._acc[con_idx] += csd_xy
-
-
-class _CohEst(_CohEstBase):
-    """Coherence Estimator."""
-
-    name = 'Coherence'
-
-    def compute_con(self, con_idx, n_epochs, psd_xx, psd_yy):  # lgtm
-        """Compute final con. score for some connections."""
-        if self.con_scores is None:
-            self.con_scores = np.zeros(self.csd_shape)
-        csd_mean = self._acc[con_idx] / n_epochs
-        self.con_scores[con_idx] = np.abs(csd_mean) / np.sqrt(psd_xx * psd_yy)
-
-
-class _CohyEst(_CohEstBase):
-    """Coherency Estimator."""
-
-    name = 'Coherency'
-
-    def compute_con(self, con_idx, n_epochs, psd_xx, psd_yy):  # lgtm
-        """Compute final con. score for some connections."""
-        if self.con_scores is None:
-            self.con_scores = np.zeros(self.csd_shape,
-                                       dtype=np.complex128)
-        csd_mean = self._acc[con_idx] / n_epochs
-        self.con_scores[con_idx] = csd_mean / np.sqrt(psd_xx * psd_yy)
-
-
-class _ImCohEst(_CohEstBase):
-    """Imaginary Coherence Estimator."""
-
-    name = 'Imaginary Coherence'
-
-    def compute_con(self, con_idx, n_epochs, psd_xx, psd_yy):  # lgtm
-        """Compute final con. score for some connections."""
-        if self.con_scores is None:
-            self.con_scores = np.zeros(self.csd_shape)
-        csd_mean = self._acc[con_idx] / n_epochs
-        self.con_scores[con_idx] = np.imag(csd_mean) / np.sqrt(psd_xx * psd_yy)
-
-
-class _PLVEst(_EpochMeanConEstBase):
-    """PLV Estimator."""
-
-    name = 'PLV'
-    accumulate_psd = False
-
-    def __init__(self, n_cons, n_freqs, n_times):
-        super(_PLVEst, self).__init__(n_cons, n_freqs, n_times)
-
-        # allocate accumulator
-        self._acc = np.zeros(self.csd_shape, dtype=np.complex128)
-
-    def accumulate(self, con_idx, csd_xy):
-        """Accumulate some connections."""
-        self._acc[con_idx] += csd_xy / np.abs(csd_xy)
-
-    def compute_con(self, con_idx, n_epochs):
-        """Compute final con. score for some connections."""
-        if self.con_scores is None:
-            self.con_scores = np.zeros(self.csd_shape)
-        plv = np.abs(self._acc / n_epochs)
-        self.con_scores[con_idx] = plv
-
-
-class _ciPLVEst(_EpochMeanConEstBase):
-    """corrected imaginary PLV Estimator."""
-
-    name = 'ciPLV'
-    accumulate_psd = False
-
-    def __init__(self, n_cons, n_freqs, n_times):
-        super(_ciPLVEst, self).__init__(n_cons, n_freqs, n_times)
-
-        # allocate accumulator
-        self._acc = np.zeros(self.csd_shape, dtype=np.complex128)
-
-    def accumulate(self, con_idx, csd_xy):
-        """Accumulate some connections."""
-        self._acc[con_idx] += csd_xy / np.abs(csd_xy)
-
-    def compute_con(self, con_idx, n_epochs):
-        """Compute final con. score for some connections."""
-        if self.con_scores is None:
-            self.con_scores = np.zeros(self.csd_shape)
-        imag_plv = np.abs(np.imag(self._acc)) / n_epochs
-        real_plv = np.real(self._acc) / n_epochs
-        real_plv = np.clip(real_plv, -1, 1)  # bounded from -1 to 1
-        mask = (np.abs(real_plv) == 1)  # avoid division by 0
-        real_plv[mask] = 0
-        corrected_imag_plv = imag_plv / np.sqrt(1 - real_plv ** 2)
-        self.con_scores[con_idx] = corrected_imag_plv
-
-
-class _PLIEst(_EpochMeanConEstBase):
-    """PLI Estimator."""
-
-    name = 'PLI'
-    accumulate_psd = False
-
-    def __init__(self, n_cons, n_freqs, n_times):
-        super(_PLIEst, self).__init__(n_cons, n_freqs, n_times)
-
-        # allocate accumulator
-        self._acc = np.zeros(self.csd_shape)
-
-    def accumulate(self, con_idx, csd_xy):
-        """Accumulate some connections."""
-        self._acc[con_idx] += np.sign(np.imag(csd_xy))
-
-    def compute_con(self, con_idx, n_epochs):
-        """Compute final con. score for some connections."""
-        if self.con_scores is None:
-            self.con_scores = np.zeros(self.csd_shape)
-        pli_mean = self._acc[con_idx] / n_epochs
-        self.con_scores[con_idx] = np.abs(pli_mean)
-
-
-class _PLIUnbiasedEst(_PLIEst):
-    """Unbiased PLI Square Estimator."""
-
-    name = 'Unbiased PLI Square'
-    accumulate_psd = False
-
-    def compute_con(self, con_idx, n_epochs):
-        """Compute final con. score for some connections."""
-        if self.con_scores is None:
-            self.con_scores = np.zeros(self.csd_shape)
-        pli_mean = self._acc[con_idx] / n_epochs
-
-        # See Vinck paper Eq. (30)
-        con = (n_epochs * pli_mean ** 2 - 1) / (n_epochs - 1)
-
-        self.con_scores[con_idx] = con
-
-
-class _DPLIEst(_EpochMeanConEstBase):
-    """DPLI Estimator."""
-
-    name = 'DPLI'
-    accumulate_psd = False
-
-    def __init__(self, n_cons, n_freqs, n_times):
-        super(_DPLIEst, self).__init__(n_cons, n_freqs, n_times)
-
-        # allocate accumulator
-        self._acc = np.zeros(self.csd_shape)
-
-    def accumulate(self, con_idx, csd_xy):
-        """Accumulate some connections."""
-        self._acc[con_idx] += np.heaviside(np.imag(csd_xy), 0.5)
-
-    def compute_con(self, con_idx, n_epochs):
-        """Compute final con. score for some connections."""
-        if self.con_scores is None:
-            self.con_scores = np.zeros(self.csd_shape)
-
-        con = self._acc[con_idx] / n_epochs
-
-        self.con_scores[con_idx] = con
-
-
-class _WPLIEst(_EpochMeanConEstBase):
-    """WPLI Estimator."""
-
-    name = 'WPLI'
-    accumulate_psd = False
-
-    def __init__(self, n_cons, n_freqs, n_times):
-        super(_WPLIEst, self).__init__(n_cons, n_freqs, n_times)
-
-        # store  both imag(csd) and abs(imag(csd))
-        acc_shape = (2,) + self.csd_shape
-        self._acc = np.zeros(acc_shape)
-
-    def accumulate(self, con_idx, csd_xy):
-        """Accumulate some connections."""
-        im_csd = np.imag(csd_xy)
-        self._acc[0, con_idx] += im_csd
-        self._acc[1, con_idx] += np.abs(im_csd)
-
-    def compute_con(self, con_idx, n_epochs):
-        """Compute final con. score for some connections."""
-        if self.con_scores is None:
-            self.con_scores = np.zeros(self.csd_shape)
-
-        num = np.abs(self._acc[0, con_idx])
-        denom = self._acc[1, con_idx]
-
-        # handle zeros in denominator
-        z_denom = np.where(denom == 0.)
-        denom[z_denom] = 1.
-
-        con = num / denom
-
-        # where we had zeros in denominator, we set con to zero
-        con[z_denom] = 0.
-
-        self.con_scores[con_idx] = con
-
-
-class _WPLIDebiasedEst(_EpochMeanConEstBase):
-    """Debiased WPLI Square Estimator."""
-
-    name = 'Debiased WPLI Square'
-    accumulate_psd = False
-
-    def __init__(self, n_cons, n_freqs, n_times):
-        super(_WPLIDebiasedEst, self).__init__(n_cons, n_freqs, n_times)
-        # store imag(csd), abs(imag(csd)), imag(csd)^2
-        acc_shape = (3,) + self.csd_shape
-        self._acc = np.zeros(acc_shape)
-
-    def accumulate(self, con_idx, csd_xy):
-        """Accumulate some connections."""
-        im_csd = np.imag(csd_xy)
-        self._acc[0, con_idx] += im_csd
-        self._acc[1, con_idx] += np.abs(im_csd)
-        self._acc[2, con_idx] += im_csd ** 2
-
-    def compute_con(self, con_idx, n_epochs):
-        """Compute final con. score for some connections."""
-        if self.con_scores is None:
-            self.con_scores = np.zeros(self.csd_shape)
-
-        # note: we use the trick from fieldtrip to compute the
-        # the estimate over all pairwise epoch combinations
-        sum_im_csd = self._acc[0, con_idx]
-        sum_abs_im_csd = self._acc[1, con_idx]
-        sum_sq_im_csd = self._acc[2, con_idx]
-
-        denom = sum_abs_im_csd ** 2 - sum_sq_im_csd
-
-        # handle zeros in denominator
-        z_denom = np.where(denom == 0.)
-        denom[z_denom] = 1.
-
-        con = (sum_im_csd ** 2 - sum_sq_im_csd) / denom
-
-        # where we had zeros in denominator, we set con to zero
-        con[z_denom] = 0.
-
-        self.con_scores[con_idx] = con
-
-
-class _PPCEst(_EpochMeanConEstBase):
-    """Pairwise Phase Consistency (PPC) Estimator."""
-
-    name = 'PPC'
-    accumulate_psd = False
-
-    def __init__(self, n_cons, n_freqs, n_times):
-        super(_PPCEst, self).__init__(n_cons, n_freqs, n_times)
-
-        # store csd / abs(csd)
-        self._acc = np.zeros(self.csd_shape, dtype=np.complex128)
-
-    def accumulate(self, con_idx, csd_xy):
-        """Accumulate some connections."""
-        denom = np.abs(csd_xy)
-        z_denom = np.where(denom == 0.)
-        denom[z_denom] = 1.
-        this_acc = csd_xy / denom
-        this_acc[z_denom] = 0.  # handle division by zero
-
-        self._acc[con_idx] += this_acc
-
-    def compute_con(self, con_idx, n_epochs):
-        """Compute final con. score for some connections."""
-        if self.con_scores is None:
-            self.con_scores = np.zeros(self.csd_shape)
-
-        # note: we use the trick from fieldtrip to compute the
-        # the estimate over all pairwise epoch combinations
-        con = ((self._acc[con_idx] * np.conj(self._acc[con_idx]) - n_epochs) /
-               (n_epochs * (n_epochs - 1.)))
-
-        self.con_scores[con_idx] = np.real(con)
-
-
-###############################################################################
-
-
-# map names to estimator types
-_CON_METHOD_MAP = {'coh': _CohEst, 'cohy': _CohyEst, 'imcoh': _ImCohEst,
-                   'plv': _PLVEst, 'ciplv': _ciPLVEst, 'ppc': _PPCEst,
-                   'pli': _PLIEst, 'pli2_unbiased': _PLIUnbiasedEst,
-                   'dpli': _DPLIEst, 'wpli': _WPLIEst,
-                   'wpli2_debiased': _WPLIDebiasedEst}
-
-
-@ verbose
-@ fill_doc
-def spectral_connectivity_epochs(
-    data, names=None, method='coh', indices=None, sfreq=None,
-    mode='multitaper', fmin=None, fmax=np.inf, fskip=0, faverage=False,
-    tmin=None, tmax=None, mt_bandwidth=None, mt_adaptive=False,
-    mt_low_bias=True, cwt_freqs=None, cwt_n_cycles=7, block_size=1000,
-    n_jobs=1, verbose=None
-):
-    """Compute bivariate (time-)frequency-domain connectivity measures.
-
-    The connectivity method(s) are specified using the "method" parameter.
-    All methods are based on estimates of the cross- and power spectral
-    densities (CSD/PSD) Sxy and Sxx, Syy.
-
-    Parameters
-    ----------
-    data : array-like, shape=(n_epochs, n_signals, n_times) | Epochs
-        The data from which to compute connectivity. Note that it is also
-        possible to combine multiple signals by providing a list of tuples,
-        e.g., data = [(arr_0, stc_0), (arr_1, stc_1), (arr_2, stc_2)],
-        corresponds to 3 epochs, and arr_* could be an array with the same
-        number of time points as stc_*. The array-like object can also
-        be a list/generator of array, shape =(n_signals, n_times),
-        or a list/generator of SourceEstimate or VolSourceEstimate objects.
-    %(names)s
-    method : str | list of str
-        Connectivity measure(s) to compute. These can be ``['coh', 'cohy',
-        'imcoh', 'plv', 'ciplv', 'ppc', 'pli', 'dpli', 'wpli',
-        'wpli2_debiased']``.
-    indices : tuple of array | None
-        Two arrays with indices of connections for which to compute
-        connectivity. Each array for the seeds and targets should contain the
-        channel indices for each bivariate connection. If ``None``, connections
-        between all channels are computed.
-    sfreq : float
-        The sampling frequency. Required if data is not
-        :class:`Epochs <mne.Epochs>`.
-    mode : str
-        Spectrum estimation mode can be either: 'multitaper', 'fourier', or
-        'cwt_morlet'.
-    fmin : float | tuple of float
-        The lower frequency of interest. Multiple bands are defined using
-        a tuple, e.g., (8., 20.) for two bands with 8Hz and 20Hz lower freq.
-    fmax : float | tuple of float
-        The upper frequency of interest. Multiple bands are dedined using
-        a tuple, e.g. (13., 30.) for two band with 13Hz and 30Hz upper freq.
-    fskip : int
-        Omit every "(fskip + 1)-th" frequency bin to decimate in frequency
-        domain.
-    faverage : bool
-        Average connectivity scores for each frequency band. If True,
-        the output freqs will be a list with arrays of the frequencies
-        that were averaged.
-    tmin : float | None
-        Time to start connectivity estimation. Note: when "data" is an array,
-        the first sample is assumed to be at time 0. For other types
-        (Epochs, etc.), the time information contained in the object is used
-        to compute the time indices.
-    tmax : float | None
-        Time to end connectivity estimation. Note: when "data" is an array,
-        the first sample is assumed to be at time 0. For other types
-        (Epochs, etc.), the time information contained in the object is used
-        to compute the time indices.
-    mt_bandwidth : float | None
-        The bandwidth of the multitaper windowing function in Hz.
-        Only used in 'multitaper' mode.
-    mt_adaptive : bool
-        Use adaptive weights to combine the tapered spectra into PSD.
-        Only used in 'multitaper' mode.
-    mt_low_bias : bool
-        Only use tapers with more than 90 percent spectral concentration
-        within bandwidth. Only used in 'multitaper' mode.
-    cwt_freqs : array
-        Array of frequencies of interest. Only used in 'cwt_morlet' mode.
-    cwt_n_cycles : float | array of float
-        Number of cycles. Fixed number or one per frequency. Only used in
-        'cwt_morlet' mode.
-    block_size : int
-        How many connections to compute at once (higher numbers are faster
-        but require more memory).
-    n_jobs : int
-        How many samples to process in parallel.
-    %(verbose)s
-
-    Returns
-    -------
-    con : array | list of array
-        Computed connectivity measure(s). Either an instance of
-        ``SpectralConnectivity`` or ``SpectroTemporalConnectivity``.
-        The shape of the connectivity result will be:
-
-        - ``(n_cons, n_freqs)`` for multitaper or fourier modes
-        - ``(n_cons, n_freqs, n_times)`` for cwt_morlet mode
-        - ``n_cons = n_signals ** 2`` with ``indices=None``
-        - ``n_cons = len(indices[0])`` when indices is supplied.
-
-    See Also
-    --------
-    mne_connectivity.spectral_connectivity_epochs_multivariate
-    mne_connectivity.spectral_connectivity_time
-    mne_connectivity.SpectralConnectivity
-    mne_connectivity.SpectroTemporalConnectivity
-
-    Notes
-    -----
-    Please note that the interpretation of the measures in this function
-    depends on the data and underlying assumptions and does not necessarily
-    reflect a causal relationship between brain regions.
-
-    These measures are not to be interpreted over time. Each Epoch passed into
-    the dataset is interpreted as an independent sample of the same
-    connectivity structure. Within each Epoch, it is assumed that the spectral
-    measure is stationary. The spectral measures implemented in this function
-    are computed across Epochs. **Thus, spectral measures computed with only
-    one Epoch will result in errorful values and spectral measures computed
-    with few Epochs will be unreliable.** Please see
-    ``spectral_connectivity_time`` for time-resolved connectivity estimation.
-
-    The spectral densities can be estimated using a multitaper method with
-    digital prolate spheroidal sequence (DPSS) windows, a discrete Fourier
-    transform with Hanning windows, or a continuous wavelet transform using
-    Morlet wavelets. The spectral estimation mode is specified using the
-    "mode" parameter.
-
-    By default, the connectivity between all signals is computed (only
-    connections corresponding to the lower-triangular part of the connectivity
-    matrix). If one is only interested in the connectivity between some
-    signals, the "indices" parameter can be used. For example, to compute the
-    connectivity between the signal with index 0 and signals "2, 3, 4" (a total
-    of 3 connections) one can use the following::
-
-        indices = (np.array([0, 0, 0]),    # row indices
-                   np.array([2, 3, 4]))    # col indices
-
-        con = spectral_connectivity_epochs(data, method='coh',
-                                           indices=indices, ...)
-
-    In this case con.get_data().shape = (3, n_freqs). The connectivity scores
-    are in the same order as defined indices.
-
-    **Supported Connectivity Measures**
-
-    The connectivity method(s) is specified using the "method" parameter. The
-    following methods are supported (note: ``E[]`` denotes average over
-    epochs). Multiple measures can be computed at once by using a list/tuple,
-    e.g., ``['coh', 'pli']`` to compute coherence and PLI.
-
-        'coh' : Coherence given by::
-
-                     | E[Sxy] |
-            C = ---------------------
-                sqrt(E[Sxx] * E[Syy])
-
-        'cohy' : Coherency given by::
-
-                       E[Sxy]
-            C = ---------------------
-                sqrt(E[Sxx] * E[Syy])
-
-        'imcoh' : Imaginary coherence :footcite:`NolteEtAl2004` given by::
-
-                      Im(E[Sxy])
-            C = ----------------------
-                sqrt(E[Sxx] * E[Syy])
-
-        'plv' : Phase-Locking Value (PLV) :footcite:`LachauxEtAl1999` given
-        by::
-
-            PLV = |E[Sxy/|Sxy|]|
-
-        'ciplv' : corrected imaginary PLV (ciPLV)
-        :footcite:`BrunaEtAl2018` given by::
-
-                             |E[Im(Sxy/|Sxy|)]|
-            ciPLV = ------------------------------------
-                     sqrt(1 - |E[real(Sxy/|Sxy|)]| ** 2)
-
-        'ppc' : Pairwise Phase Consistency (PPC), an unbiased estimator
-        of squared PLV :footcite:`VinckEtAl2010`.
-
-        'pli' : Phase Lag Index (PLI) :footcite:`StamEtAl2007` given by::
-
-            PLI = |E[sign(Im(Sxy))]|
-
-        'pli2_unbiased' : Unbiased estimator of squared PLI
-        :footcite:`VinckEtAl2011`.
-
-        'dpli' : Directed Phase Lag Index (DPLI) :footcite:`StamEtAl2012`
-        given by (where H is the Heaviside function)::
-
-            DPLI = E[H(Im(Sxy))]
-
-        'wpli' : Weighted Phase Lag Index (WPLI) :footcite:`VinckEtAl2011`
-        given by::
-
-                      |E[Im(Sxy)]|
-            WPLI = ------------------
-                      E[|Im(Sxy)|]
-
-        'wpli2_debiased' : Debiased estimator of squared WPLI
-        :footcite:`VinckEtAl2011`.
-
-    References
-    ----------
-    .. footbibliography::
-    """
-    (
-        fmin, fmax, n_bands, method, con_method_types, accumulate_psd,
-        parallel, my_epoch_spectral_connectivity
-    ) = _check_spectral_connectivity_epochs_settings(
-        method, fmin, fmax, n_jobs, verbose, _CON_METHOD_MAP)
-
-    (names, times_in, sfreq, events, event_id,
-     metadata) = _check_spectral_connectivity_epochs_data(data, sfreq, names)
-
-    # loop over data; it could be a generator that returns
-    # (n_signals x n_times) arrays or SourceEstimates
-    epoch_idx = 0
-    logger.info('Connectivity computation...')
-    warn_times = True
-    for epoch_block in _get_n_epochs(data, n_jobs):
-        if epoch_idx == 0:
-            # initialize everything times and frequencies
-            (times, n_times, times_in, n_times_in, tmin_idx, tmax_idx, n_freqs,
-             freq_mask, freqs, freqs_bands, freq_idx_bands, n_signals,
-             warn_times) = _prepare_connectivity(
-                epoch_block=epoch_block, times_in=times_in, tmin=tmin,
-                tmax=tmax, fmin=fmin, fmax=fmax, sfreq=sfreq, mode=mode,
-                fskip=fskip, n_bands=n_bands, cwt_freqs=cwt_freqs,
-                faverage=faverage)
-
-            # check indices input
-            n_cons, indices_use = _check_indices(indices, n_signals)
-
-            # get the window function, wavelets, etc for different modes
-            (spectral_params, mt_adaptive, n_times_spectrum,
-             n_tapers) = _assemble_spectral_params(
-                mode=mode, n_times=n_times, mt_adaptive=mt_adaptive,
-                mt_bandwidth=mt_bandwidth, sfreq=sfreq,
-                mt_low_bias=mt_low_bias, cwt_n_cycles=cwt_n_cycles,
-                cwt_freqs=cwt_freqs, freqs=freqs, freq_mask=freq_mask)
-
-            # unique signals for which we actually need to compute CSD/PSD
-            sig_idx = np.unique(np.r_[indices_use[0], indices_use[1]])
-            n_signals_use = len(sig_idx)
-
-            # map indices to unique indices
-            idx_map = [np.searchsorted(sig_idx, ind) for ind in indices_use]
-
-            # allocate space to accumulate PSD
-            if accumulate_psd:
-                if n_times_spectrum == 0:
-                    psd_shape = (n_signals_use, n_freqs)
-                else:
-                    psd_shape = (n_signals_use, n_freqs, n_times_spectrum)
-                psd = np.zeros(psd_shape)
-            else:
-                psd = None
-
-            # create instances of the connectivity estimators
-            con_methods = []
-            for mtype in con_method_types:
-                con_methods.append(mtype(n_cons=n_cons, n_freqs=n_freqs,
-                                         n_times=n_times_spectrum))
-
-            sep = ', '
-            metrics_str = sep.join([meth.name for meth in con_methods])
-            logger.info('    the following metrics will be computed: %s'
-                        % metrics_str)
-
-        call_params = dict(
-            sig_idx=sig_idx, tmin_idx=tmin_idx, tmax_idx=tmax_idx, sfreq=sfreq,
-            method=method, mode=mode, freq_mask=freq_mask, idx_map=idx_map,
-            n_cons=n_cons, block_size=block_size,
-            psd=psd, accumulate_psd=accumulate_psd,
-            mt_adaptive=mt_adaptive,
-            con_method_types=con_method_types,
-            con_methods=con_methods if n_jobs == 1 else None,
-            n_signals=n_signals, n_signals_use=n_signals_use, n_times=n_times,
-            gc_n_lags=None, multivariate_con=False,
-            accumulate_inplace=True if n_jobs == 1 else False)
-        call_params.update(**spectral_params)
-
-        epoch_idx = _compute_spectral_methods_epochs(
-            con_methods, epoch_block, epoch_idx, call_params, parallel,
-            my_epoch_spectral_connectivity, n_jobs, n_times_in, times_in,
-            warn_times)
-
-    # normalize
-    n_epochs = epoch_idx
-    if accumulate_psd:
-        psd /= n_epochs
-
-    # compute final connectivity scores
-    con = list()
-    for conn_method in con_methods:
-
-        # future estimators will need to be handled here
-        if conn_method.accumulate_psd:
-            # compute scores block-wise to save memory
-            for i in range(0, n_cons, block_size):
-                con_idx = slice(i, i + block_size)
-                psd_xx = psd[idx_map[0][con_idx]]
-                psd_yy = psd[idx_map[1][con_idx]]
-                conn_method.compute_con(con_idx, n_epochs, psd_xx, psd_yy)
-        else:
-            # compute all scores at once
-            conn_method.compute_con(slice(0, n_cons), n_epochs)
-
-        # get the connectivity scores
-        this_con = conn_method.con_scores
-
-        if this_con.shape[0] != n_cons:
-            raise RuntimeError(
-                'first dimension of connectivity scores does not match the '
-                'number of connections; please contact the mne-connectivity '
-                'developers')
-        if faverage:
-            if this_con.shape[1] != n_freqs:
-                raise RuntimeError(
-                    'second dimension of connectivity scores does not match '
-                    'the number of frequencies; please contact the '
-                    'mne-connectivity developers')
-            con_shape = (n_cons, n_bands) + this_con.shape[2:]
-            this_con_bands = np.empty(con_shape, dtype=this_con.dtype)
-            for band_idx in range(n_bands):
-                this_con_bands[:, band_idx] = np.mean(
-                    this_con[:, freq_idx_bands[band_idx]], axis=1)
-            this_con = this_con_bands
-
-        con.append(this_con)
-    # No patterns for bivariate connectivity
-    patterns = [None for _ in range(len(con))]
-
-    # return all-to-all connectivity matrices raveled into a 1D array
-    if indices is None:
-        logger.info('    assembling connectivity matrix')
-        con_flat = con
-        con = list()
-        for this_con_flat in con_flat:
-            this_con = np.zeros((n_signals, n_signals) +
-                                this_con_flat.shape[1:],
-                                dtype=this_con_flat.dtype)
-            this_con[indices_use] = this_con_flat
-
-            # ravel 2D connectivity into a 1D array
-            # while keeping other dimensions
-            this_con = this_con.reshape((n_signals ** 2,) +
-                                        this_con_flat.shape[1:])
-            con.append(this_con)
-
-    conn_list = _store_results(
-        con=con, patterns=patterns, method=method, freqs=freqs,
-        faverage=faverage, freqs_bands=freqs_bands, names=names, mode=mode,
-        indices=indices, n_epochs=n_epochs, times=times, n_tapers=n_tapers,
-        metadata=metadata, events=events, event_id=event_id, rank=None,
-        gc_n_lags=None, n_signals=n_signals)
-
-    return conn_list
